Use CFRetain() to ensure the CFRunLoop! is kept alive as long as we need it
